Each packet now tracking the index of last entry
The Packet object now has a new field which keeps track of the index of
the entry returned by the last match table lookup. There is a special
value if the last lookup was a miss. Unfortunately, this means that the
index part of the entry handle is no longer fully hidden by the
implementation, which I think was nice. This index should be ignored in
most use cases; however it is needed for the implementation of
direct-like externs.