Add ability to subscribe to group member updates
Target code can now subscribe to notifications signalling the addition /
removal of a member to / from a group in a indirect table.

Not sure if this is the best way of doing this, or if this change is
here to stay, but that was the easiest and less intrusive way I could
think of.