Add ability to dump first packet bytes in log
This is done using the --dump-packet-data command line option. This
option takes one argument (the number of bytes to dump). Bytes are
dumped when the packet is received and when its is sent. Because at this
stage no Packet instance exists yet, the packet id cannot be included in
the log message.

This commit also includes some minor improvements and fixes a casting
bug.