Add support for string type for primitive params
Architecture-specific action primitives can now support parameters with
a string type. In the C++ implementation of the primitive, the parameter
type can be either 'const std::string &' or 'const char *'. See
tests/test_actions.cpp for an example.

The JSON documentation was updated to reflect this addition, and the
version number was bumped up to 2.2.