Add support for P4_16 enum declarations
This is very similar to how we handle errors (see #259). It is expected
that the bmv2 compiler will assign a unique numerical value to each
enum constant. This mapping (enum name to value) will be made
available in the bmv2 JSON input. In the rest of the JSON file, the
enum constants are referenced only through their numerical value.

The JSON documentation was updated to include enum support and the JSON
version was bumped up to 2.3.