Allow switch to start without JSON configuration
The target binaries can be started without an input JSON config by using
the --no-p4 command line flag. Until a config has been pushed to the
switch however, no traffic can be processed. The start_and_return method
was modified to block until a config is pushed. The virtual method that
needs to be implemented by switches was renamed
start_and_return_. Similarly the receive virtual method was also renamed
receive_. This means that existing target switches outside of this repo
will need to be updated.

The tools (e.g. runtime CLI) needed to be updated as a result to support
the empty config case.