Add tentative support for method calls in parser
We add support for calling primitive actions (in bmv2 terminology that
includes extern methods) in parse states. This is a tentative change
with a tentative JSON format, which is not documented in JSON_format.md
yet.

I re-organized some P4Objects code to avoid code duplication in
primitive initialization between the action code and the parser code.