Add a exact match cache for ternary + range tables
With the current implementation a ternary lookup often incurs a linear
search over the whole table. We add a small exact match cache (by
default 64 entries) to each ternary / range table. Assuming that many
successive packets actually belong to the same "flow" and hit the same
ternary entry, this should considerably speed-up the average lookup
time. We use LRU for cache eviction.

The cache is only used when more than 16 entries have been added to a
table, to avoid any kind of penalty in the case where the linear search
is actually pretty small.

We chose to be very conservative for cache invalidation. Every time an
entry is added or removed to the table, the cache is emptied. This can
be improved in the future.