Add support for P4_16 header unions
There is a P4_16 proposal to add header unions to the language. This
patch adds support for them in bmv2 and defines the appropriate JSON
input format when using header unions. Having first-class support for
unions in bmv2 allows for easier debugging, with potentially better log
messages.

The most important use of header unions is when using stacks / arrays of
them. This enables writing programs able to serialize / deserialize
TLV-style protocol options (e.g. IPv4 options). Which is why this patch
adds support for stacks of header unions, which actually took the
biggest effort. Header unions and header union stacks are modeled after
header stacks.

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.10