Remove clang-3.6 from Travis regressions
It does not seem to be compatible with recent (6.2) libstdc++, which is
causing issues when compiling. There are no such issues with
clang-3.8. I think it is more important to be able to run the
regressions with gcc6, and I don't want to spend too much time trying to
find a workaround. I may add another (more recent) version of clang to
the list of compilers used in the regression in the future.