Add support for conditional jump in action bodies
We tentatively add support for 2 "core" action primitives for absolute
jumps: `_jump` and `_jump_if_zero`. `_jump` takes one parameter which
must resolve to an integral value `offset`. When it is executed, we jump
to the primitive call at index `offset` in the enclosing
action. `_jump_if_zero` is a conditional jump and takes two parameters
which must both resolve to integral values, which we call respectively
`cond` and `offset`. If `cond` is 0, the primitive behaves exactly like
`_jump` and we jump to the primitive call at index `offset`; otherwise
the primitive is a no-op.

This was implemented at no additional cost compared to the existing
implementation. In other words, if an input does not use these 2
primitives, performance will not be affected.

This can be used by a P4_16 compiler to support conditions in actions.

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.13