Fix issue in entry lookup for ternary match tables
When adding 2 entries to the table, then removing the first added; we
would observe an error when sending a packet matching the second entry
added only (bmv2 would report a table hit instead of a miss, with a
corrupt entry handle).

This was caused by some incorrect code to compute an entry handle based
on the location of the entry in the table "array".