Add packet IO support to simple_switch_grpc
New target-specific command line option for simple_switch_grpc:
--cpu-port, to specify a CPU port number. Packet-out traffic will be
injected on that port and traffic received on that port will be sent
through p4runtime as packet-out traffic.
This port number is special and is not known to bmv2 DeviceMgr code; do
not add an interface with this port number (using --interface / -i).

A test program was added (test_packet_io), but we need a better
infrastructure for simple_switch_grpc unit tests.