Support for injecting / receiving pkts with gRPC
We define a new p4::bm::DataplaneInterface gRPC service which can be
used to inject and receive packets. We add a DevMgrIface implementation
to simple_switch_grpc which handles this gRPC
service. simple_switch_grpc users can choose to use this DevMgrIface
implementation by providing a gRPC server address with the new
`--dp-grpc-server-addr` command-line option.

In this commit, we also start using GTest for unit-testing
simple_switch_grpc. For now we only have tests for the new gRPC packet
injection / reception feature. The legacy tests will be converted to
GTest in a future commit.