Dissociate checksum verification and update
One can now have checksums that are just verified (in all parsers),
checksums that are just updated (in all deparsers) and checksums that
are both verified and updated. This can be specified in the JSON using
the new 'update' and 'verify' attributes for checksums.

This change is backward-compatible.

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.14.