Initial support for gNMI & OpenConfig
This leverages PI support for gNMI and OpenConfig, so you need a recent
version of the PI. Just like for the PI project, sysrepo needs to be
installed and you need to run configure (for simple_switch_grpc) with
--with-sysrepo.

The code registers a subscriber for config updates and an operational
state provider with sysrepo. As of now we do not react to config updates
as we only seek to support gNMI Subscribe requests. We still assume that
dataplane ports are configured through an external mechanism (most
likely, through the command-line when starting simple_switch_grpc).