Build simple_switch_grpc in Dockerfile
And run `make check` for simple_switch_grpc in Travis.
This change means that the bmv2 Docker image now "inherits" from the PI
one, and not the other way around. New docker files had to be introduced
for both repos to avoid a circular dependency, since PI uses bmv2 for
some CI testing.

Because of C++ ABI incompatibilities (my best guess), we only compile
and run tests for simple_switch_grpc when the compiler is gcc which is a
pretty ugly hack.