Show node name in log when evaluating conditions
I have found that there are P4 programs with 'if' statements inside of
controls that are called multiple times from ingress or egress, where
the p4c-bm2-ss inlining it performs causes multiple condition nodes to
be created with exactly the same source info.

I have a use case where distinguishing exactly which node is being
evaluated is useful to know.