Support providing match key element name in JSON
Introduced a new optional JSON attribute for match key fields,
'name'. In P4_16 a table can match on arbitary expressions and while the
compiler can evaluate these expressions upstream and match on a
temporary field holding the expression result, it is desirable to
preserve the orginal name.

As far as bmv2 itself is concerned, this information is only useful for
debugging, but this onformation could be required by control-planes.

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.15

Fixes #500