Implement P4_16 conformant stack support
For header stacks and header union stacks. We support both behaviors at
the moment (legacy and P4_16). The legacy implementation is the default
one and will be the default for the 1.10 release. The P4_16 conformant
behavior can be toggled by using the `--enable-WP4-16-stacks` configure
flag. The plan is to have the new implementation be the default after
the 1.10 release.

The new implementation differs from the legacy one as follows:
push_front and pop_front shift the entire stack (not just until the next
index), and pushed headers are marked as invalid instead of valid.

We have tried to minimize the amount of code guarded by ifdef
directives. We used the CRTP to implement both behaviors and both
implementations are always compiled. We have unit tests for both
implementations as well.