Fix simple_switch_grpc direct counter GTest
Because PSA mandates that a given table can have at most one of each
direct resource, we have remove the id from the DirectCounterEntry &
DirectMeterEntry messages in P4Runtime (the id can be inferred from the
table id, included in the TableEntry field). As a result one of the
simple_switch_grpc unit tests had to be updated to fix a compialtion
issue.