Fix extraction function for variable-length fields
For VL fields, the bitwidth is dynamic and only known at extract
time. Which means that extract_VL is in charge of computing the Bignum
mask for the Field, which is then used to ensure that the field stays in
the correct range. However, we were using an incorrect expression to
compute the mask, which was causing it to be always 0 for bitwidths >=
32.

Fixes issue #538