Accept unique suffixes as valid identifiers in CLI
A recent change in p4lang/p4c added the missing top-level control name
to the fully-qualified name of objects defined within this control. The
fully-qualified is what appears in the bmv2 JSON and what is expected by
the Thrift server. To avoid breaking exsiting examples / tutorials which
use the bmv2 runtime_CLI to install table entries, we now accept unique
suffixes of the fully-qualified name as valid indentifiers in the
runtime_CLI.

New code should probably transition to the correct fully-qualified name
instead of relying on this feature. Which is why auto-completion does
not suggest suffixes.