Support OC interface counters in sysrepo state provider
We now support openconfig-interfaces counters in the sysrepo state
provider for simple_switch_grpc.

In order to achieve this, we updated the bm::DevMgrIface to support
querying and clearing port stats (packets & octets). We also considered
implementing this functionality in bm::DevMgr, but decided it was better
to let each DevMgrIface implementation support stats in its preferred
way.

Both the BMI DevMgr and the DataplaneInterface gRPC service (for
simple_switch_grpc) support stats. For BMI, the implementation is quite
naive and uses a mutex to protect the stat counters. This may be
improved in the future with CAS.

The simple_switch_grpc README has been updated to indicate the
additional OpenConfig support.