Fix simple_switch_grpc transmit fn
Because of a recent change, simple_switch_grpc would crash upon
transmitting a packet except when using the gRPC dataplane service. This
fix modifies the transmit function to treat outgoing packets differently
based on whether the gRPC dataplane service is used (in which case the
packet id needs to be passed along) or some other DevMgrIface
implementation is used.

Fixes #582