Add option --load-modules to simple_switch
Add option --load-modules to allow loading external code as .so files.
This allows externs to be registered, e.g. using REGISTER_PRIMITIVE or
BM_REGISTER_EXTERN_METHOD from external code, without changing
behavioral-model's code.

The .so files are comma-delimited, e.g. -- --load-modules
file1.so,file2.so .

This feature is enabled automatically in compile time if dlopen is found.
To disable it, add --disable-modules to ./configure's parameters.