Use instance type values specified in JSON
Instead of hard-coding integer values for different packet instance
types and requiring P4 programmers to be aware of this target-dependent
mapping, we now support defining the different instance types in the P4
architecture file as an enum. Simple switch will use the integer values
(or the enum members) assigned by the compiler backend.

This commit assumes that the enum will be called "InstanceType" in
v1model.p4 and also hard-codes the enum member names. We can amend the
names and update this commit if needed.

This change is backward-compatible: if the JSON does not include a value
for a specific enum member, bmv2 will print a warning and use a default
value, which corresponds to the legacy value used by previous
simple_switch versions. Ideally the check would be slightly different:
we would check if the enum name exists and if all the required members
are present; if something is missing we would use the default values for
ALL members instead of defaulting on a per member basis. But this would
be a little harder to implement.