Add support for sending digests through P4Runtime
For the simple_switch_grpc target. This is done by registering a custom
TransportIface implementation with the Switch base class, to capture all
the notifications which are usually broadcast on a Nanomsg PUBSUB
socket.

The P4Runtime implementation in p4lang/PI requires that no digest
notification be sent while a forwarding pipeline config swap is
ongoing. To ensure that we respect this assumption, we now generate a
new kind of notifications for "swap" events. These notifications are
captured by our custom TransportIface implementation for
simple_switch_grpc, to ensure that all digest notifications are ignored
during the swap.

The only "shortcut" is that ideally simple_switch_grpc should not be
generating learning notifications if the digest has not been configured
yet through P4Runtime or if learning was disabled for this digest
through P4Runtime. At the moment we rely on the P4Runtime implementation
to drop the notifications sent through PI.