Use config.h file to fix use of #ifdef in headers
Ideally we would not use custom #ifdef in headers at all but this is a
bigger project for the future...

the bmv2 library with BMDEBUG_ON defined but within a separate project
(simple_switch_grpc) without BMDEBUG_ON defined, which was causing the
bm::Field class to have a different layout in the library and in the
application. This was causing different segfaults based on the other
preprocessor flags and compiler flags used and was difficult to diagnose
(wasn't caught by Valgrind).

To avoid this issue, we now rely on the config.h file generated from
configure.ac. This file needs to be installed because it is required by
installed bmv2 headers. To avoid interfering with other libraries, we
use AX_PREFIX_CONFIG_H to install config.h under @includedir@/bm and to
prefix all defined macros with BM_.

Fixes #715