New 'advance' parser operation
This new operation is more powerful than 'shift', which can only shift
by a fixed amount of bytes. 'advance' takes as a parameter either an
hexadecimal string ('hexstr'), a field ('field'), or an arithmetic
expression ('expression').

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.19.

Fixes #727