Support for arbitrary uint32 port numbers
Interfaces can now be bound to bmv2 with arbitrary port numbers. The
following components required changes:
 * BMI, which reads / writes packets from / to Linux interfaces using
   libpcap.
 * the various Queueing implementations, which buffer packets between
   ingress & egress in simple_switch and other targets. This was the
   bulk of the work and I'm a bit disappointed with the trade-offs I had
   to make; in particular we now have a "global" lock instead of a lock
   per worker.

BMI still requires an upper bound on the max *number* of ports (and not
on the max port number) for implementation reasons. By default this
upper bound is 512 but it can be increased with --max-port-count.

Note that for v1model.p4 (base architecture for simple_switch) the
default bitwidth for port fields is 9-bit, so only ports in the [0, 511]
range are supported. The burden is on the v1model user to change the
bitwidth if so desired. In psa.p4, the default bitwidth is 32.

Fixes #743