Implement mark_to_drop primitive for simple_switch
The primitive takes a header as a parameter, whose header type must have
a field called 'egress_spec'. This is part of an attempt to fix
mark_to_drop support in v1model / simple_switch. Because the
implementation of the primitive modifies a user-exposed field, it should
not be done in a hidden way. The primitive still modifies
intrinsic_metadata.mcast_grp instead of looking for a 'mcast_grp' field
in the provided header, but fixing this requires a bigger change (we
need to deprecate intrinsic_metadata.mcast_grp in favor of
standard_metadata.mcast_grp).