Change "private" v1model standard_metadata fields to BMv2 packet registers
Remove the use of the following 4 fields from BMv2 simple_switch and
simple_switch_grpc.  They no longer need to be in the BMv2 JSON file
for such programs, and clone, recirculate, resubmit, and digest
operations are still able to work without them there, because they are
now stored in a BMv2 "packet register", similar to how the packet
length field currently is.

+ clone_spec
+ resubmit_flag
+ recirculate_flag
+ lf_field_list