Make exit a core primitive
"exit" is a P4_16 built-in statement and should therefore be implemented
in a bmv2 target-agnostic way as much as possible. We make exit a bmv2
core primitive, and remove the exit definition from simple_switch +
psa_switch.

exit is therefore added to the JSON_format document, and the JSON
version number is bumped up to 2.21.

Note that at the moment targets are still required to call
`Packet::reset_exit` when the pipeline exits to reset the flag. I feel
like this could be done just before returning from `Pipeline::apply`,
but there may be some historical reason why I did not do it that way (so
to be safe I did not change anything...).

Fixes #790