Cleanup the "parameters_vector" code
This is the code that enables action primitives to take as parameters
lists of values (for the moment, we are restricted to lists of Data
objects).

We refactor the code to get to a cleaner design: we no longer have a
class (ActionParamVectorFn) that both inherits from and encapsulates
another one (ActionFn). This led to less duplication and fewer lines of
code.

We have removed the check in P4Objects that verified that only arguments
resolving to numerical values (i.e. that can be cast to Data) were added
to vector of parameters. In theory we can support all homogeneous lists
in bmv2, even though it is not the case yet. We should think of how we
can define a proper check in the future: one that enforces that the list
is homogeneous and the list type is supported by that version of bmv2.

Some cleanup was done in the Gtests as well: we no longer use the
console logger to verify that messages are logged properly, instead we
use a string stream.

The JSON version number was bumped up to 2.22, since it was not done in
the original "parameters_vector" patch.