Add support for parser value sets (PVS) to runtime_CLI
We now provide the following commands: pvs_add, pvs_remove, pvs_get,
pvs_clear. Implementing pvs_get and pvs_clear required adding the
equivalent RPCs to the Thrift API.

This commit also adds a bmv2 JSON sanity check for the "mask" attribute,
when it is used in a PVS parser transition.

The included JSON file, pvs_struct_2.json, was used to test the
implementation of the CLI commands. It was generated from a P4 program
written by @jafingerhut for the compiler testsuite.

Fixes #855