psa_switch: Add recirculation/resubmission limit
Add max packet recirculation and resubmission limit. By default the limit
for recirculation and resubmission is set to 255 (unlimited).

The options set_max_recirculations and set_max_resubmissions are added
to psa_switch_CLI to enable the limit to be changed to a value between
0 and 254; the special value of 255 (or -1) is used as unlimited.

In addition, get_max_recirculations and get_max_resubmissions can be
used to obtain the current recirculation/resubmission limit.

The metadata fields max_recirculations and max_resubmissions for the
types psa_ingress_input_metadata_t and psa_egress_input_metadata_t
are set to a value corresponding to the current limit for a packet,
and this value is updated after recirculation/resubmission.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>