Fix handling of non-Data expressions as primitive arguments
When adding support for returning lvalues in the expression evaluation
engine (https://github.com/p4lang/behavioral-model/pull/839), we
introduced support for using expressions which do not evaluate to a
simple Data value as primitive arguments. In particular, we can use an
expression which evaluates to a Header object as a primitive
argument. However, the code responsible for parsing the bmv2 input JSON
and instantiating Action objects was treating all expressions used as
primitive arguments as "Data expressions", thus causing bmv2 to crash
when it was actually not the case.

Fixes #887