Fix type deduction for expressions when importing JSON
A recent fix for a separate issue
(https://github.com/p4lang/behavioral-model/pull/888) exposed a separate
bug: the type of expressions for which the top-level operator was
"access_field" was not deduced properly (set to UNKNOWN instead of
DATA), which caused bmv2 to assert when building the Action object.

The same issue existed for "access_union_header", which is an operator
introduced more recently (and possibly not used by p4c yet).

Fixes #889