Fix GroupSelectionIface implementation used by PI
PI uses a custom GroupSelectionIface implementation to support watch
ports for P4Runtime action profile group members.

Before this fix, membership information in the GroupSelectionIface
implementation wasn't cleaned up properly when a group was deleted,
which means that when a new group was created with the same id (but
different membership), table lookups would yield invalid members and
bmv2 would assert.

The fix consists in making the appropriate method calls to remove each
member from the group when the group is being deleted.

Fixes #893