Fix build with recent Protobuf & gRPC versions
When testing with Protobuf 3.12 and gRPC 1.30, the build failed because
of a change in the gRPC header (a grpc::Server typedef was introduced as
a replacement for the grpc::Server class, which means our forward
declaration was no longer correct).

We also replace all grpc++ include statements with grpcpp include
statements. grpc++ headers are deprecated since gRPC 1.10 (the official
gRPC version for p4lang software is 1.17).

Fixes #925