Reset meter rates when getting default meter spec from PI
The code was incorrect as it was assuming that converting the default
meter spec to bmv2 rates and then back to a PI spec would yield the
default spec again. This is not a valid assumption because we convert
the meter rates from an integer to a float and then back to an integer.

Fixes #1029