Remove redundant moves in return statements
With recent g++ versions:

../include/bm/bm_sim/queueing.h:780:24: error: redundant move in return statement [-Werror=redundant-move]
  780 |     return std::move(fn);
      |                        ^
../include/bm/bm_sim/queueing.h:780:24: note: remove 'std::move' call
cc1plus: all warnings being treated as error

While copy elision is not possible when retruning a function parameter,
if the other conditions for NVO are met, a move operation should be
used. The std::move doesn't cause the compiler to perform worse (unlike
for RVO), but it is redundant.